requirements.txt
# --- Affect-Diff Core Dependencies ---

# Deep Learning Framework
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pytorch-lightning>=2.2.1

# Foundation Model Backbones
transformers>=4.38.0

# Generative & Architectural Utilities
einops>=0.7.0
numpy>=1.26.0
scipy>=1.10.0

# Multimodal Data Engineering (CMU-SDK)
mmsdk>=1.1.0
pickledb>=0.9.2
h5py>=3.10.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Experiment Tracking & Configuration
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.16.0

# Metrics & Logging
torchmetrics>=1.3.0
tqdm>=4.66.0

# Signal Processing & Visualization
librosa>=0.10.0
matplotlib>=3.7.0
